MNT/DEV/BUG: fix filenamerender; add paths; format
I fixed the bug in filenamerender. It cannot mutate the doc in place. Instead, it should return the name and doc and other callbacks use the returned values.

I add the features that there could be more than on tiff_bases. The data will be saved to theses tiff_bases.

I use black to format all the python files.